# Use Node pointers plus object pool, eliminate separate vertices array