@mourner 2015-10-06
> This implements an optimization I think we should try for the JS implementation: f4b0a71.

Interesting! So you suggest storing `x` and `y` in `Node` instances instead of using `data[node.i], data[node.i + 1]` for all calculations, right? 

Why would this be faster? Better memory access, or shorter path to data?