@jfirebaugh 2015-10-06
Yeah, I don't know for sure that it would be faster, but intuitively, since you usually already have a `node` reference available, it should be faster because it avoids two memory accesses and may also improve cache locality.