@jfirebaugh 2015-10-07
I would like to eliminate the `boost::pool` dependency from this PR and instead use a simple preallocated `std::unique_ptr<Node[]>` with one `Node` for each input vertex. This would depend on eliminating the second place that Nodes are created: 

https://github.com/mapbox/earcut.hpp/blob/bfbe17d74dd68207219c8c63d6338d116741799d/include/earcut.hpp#L761-L762

@mourner @kkaefer Is it somehow possible to reuse the existing Nodes in `splitPolygon` rather than creating new ones?