@mourner 2015-10-08
:+1: for merging, @kkaefer said :+1: in Slack too.